almost..... rooms....
ive gotten further but ... somehow the logic off how to add to arrays together and putting them back is kinda hard.

i have lots of ideas how to do this right though.... but there is an other issue, where when a room is found afterwards the logic for detecting new rooms fails and it always detects a new room which is wrong ... sad that there is no full break. although there is probably the possibiliy, that one block connects more then 3 rooms at the same time... so ya those kind of cases have to be accounted for too.