slopes and rooms old way again
i did some reading and looking and i found that raycasts in godot arent the best... they allocate a lot of space for each use... also i came to the conclusion that they need to work in the physic... so they need to be on that thread too... so no multithreading, which would be a deal breaker for this. so we go back to this aproach and ill just make it work...

This is just a previous state of the rooms file

inseted into the current state of the game.