Sort of Working Room Recignition
I finally made it to a working state with the Room File... Ive spent a lot of time on this. Its hard to wrap my head around for some reason. its basically deep first search but with a "touching border" recignition. If i keep struggling with it i might just go with the first attemd which was just to offset the start and end position by one to have a lay of air always around the vessel...