further progress on rooms
i changed a lot of things... first of all a bunch of orders of operations have been put in a more sensible deal...ish and also i changed core things about how surrounding air blocks are found.... also i added padding to all sizes so there should be air always connecting all the blocks... as of now, there is still an issue with that where if a volume is open to the top it dosent work yet correctly. but if a room is actually enclosed it will be recigonized correctly.