Hydraulic Erosion is sort of working
Im not really happy with how it looks but it is usable for now. Thanks to Sebastian Langue for his Video, Demo Project and sharing of this work it was a great help. Its Based on his design... again somehow i really didnt work that great comparted to his but oh well