Further Improvments to the Hydraulic Erosion
I got this to work decently well i think it looks better now.

But Ill Try a different Method layed out by a paper in 2011 where instead of simulating individual drops there is a sort of flow field that will determin where settlement is carried to.

Im hoping that it can run maybe quicker, allow for lakes and rivers and also creates an overall more plataues. Also it might be the key to the player actually having rivers flow and being able to change how they flow.... that sounds completly crazy in a lot of ways so im curious if i can make it work from a 10 page PDF file.

https://inria.hal.science/inria-00402079/document