some more minor tweaks, and thoughts about multithreading
Ive crudly tested what causes the high runtime ... i had an idea where i could have the mainthread, that runs the terrian and only calculate the path and give those coardinates into a thread that will calculate the new hights for each position and then deliver them for the mainthread act on them ... but it seems thats really the lesser part of the issure. i think the for loops only by them selfs allrady cause significant performance issues. obviously the issue scales quadratically with map size which is needed for the player to even consider making things high speed and actually use them...