Terrain Generation big leap
now by fling into a chunk and pressing T i can load the 4 regions that are poart of the chunk and have them be hydraulically eroded automatically and saved .... by pressing Z they will be loaded... the next thing for me is making it so, that the chunks to load are automatically selected. Since this is fully multithreaded. as long as regions that are used for the chunk dont overlap as many chunks can be generated as there are threads... So my next step making that automatic after that i kinda feel like adding some vegitagion... i kinda want it to be procedually generated as well.

i think is still not working perfectly were it will be fine with different region sizes it seems to be offset at 2x2 kms but not with 1x1 km